CLOSED - task #14: Get all GitLab users using GitLab API http://github.com/Kevin-Lee/trac-issues-to-gitlab/issues/issue/14
*Changed
-JDK 1.7 -> 1.6

+++ src/main/java/com/lckymn/kevin/gitlab/api/GitLabApiUtil.java
@@ -59,12 +59,12 @@ public final class GitLabApiUtil
         .toString();
   }
 
-  public static String prepareUrlForMilestones(final String url, final String privateToken, final Long projectId)
+  public static String prepareUrlForMilestones(final String url, final String privateToken, final Integer projectId)
   {
     return prepareUrl(url, privateToken, projectId, MILESTONES);
   }
 
-  public static String prepareUrlForIssues(final String url, final String privateToken, final Long projectId)
+  public static String prepareUrlForIssues(final String url, final String privateToken, final Integer projectId)
   {
     return prepareUrl(url, privateToken, projectId, ISSUES);
   }
